refactoring and godoc
