+++ pkg/transport/process/processors/sleep/main.go
@@ -11,13 +11,13 @@ import (
 	"syscall"
 	"time"
 
+	"github.com/gardener/component-cli/pkg/transport/process"
 	"github.com/gardener/component-cli/pkg/transport/process/extensions"
-	"github.com/gardener/component-cli/pkg/transport/process/processors"
 )
 
 const sleepTimeEnv = "SLEEP_TIME"
 
-// a test processor which sleeps for a configurable duration and then exists with an error.
+// a test processor which sleeps for a configurable duration and then exits with an error.
 func main() {
 	sleepTime, err := time.ParseDuration(os.Getenv(sleepTimeEnv))
 	if err != nil {
@@ -36,7 +36,7 @@ func main() {
 		log.Fatal("finished sleeping -> exit with error")
 	}
 
-	srv, err := processors.NewUDSServer(addr, h)
+	srv, err := process.NewUDSServer(addr, h)
 	if err != nil {
 		log.Fatal(err)
 	}
